MainServer/requirements.txt
fastapi
uvicorn
python-multipart
mysql-connector-python
configparser
